feat(core): add copy link button to local share menu (#9271)
close AF-1838
