+++ packages/frontend/core/src/components/affine/share-page-modal/share-menu/index.css.ts
@@ -54,6 +54,12 @@ export const copyLinkContainerStyle = style({
   alignItems: 'center',
   width: '100%',
   position: 'relative',
+  selectors: {
+    '&.secondary': {
+      padding: 0,
+      marginTop: '12px',
+    },
+  },
 });
 export const copyLinkButtonStyle = style({
   flex: 1,
@@ -64,6 +70,14 @@ export const copyLinkButtonStyle = style({
   borderBottomRightRadius: '0',
   color: 'transparent',
   position: 'initial',
+  selectors: {
+    '&.dark': {
+      backgroundColor: cssVarV2('layer/pureBlack'),
+    },
+    '&.dark::hover': {
+      backgroundColor: cssVarV2('layer/pureBlack'),
+    },
+  },
 });
 export const copyLinkLabelContainerStyle = style({
   width: '100%',
@@ -80,6 +94,11 @@ export const copyLinkLabelStyle = style({
   transform: 'translateX(-50%) translateY(-50%)',
   lineHeight: '20px',
   color: cssVarV2('text/pureWhite'),
+  selectors: {
+    '&.secondary': {
+      color: cssVarV2('text/primary'),
+    },
+  },
 });
 export const copyLinkShortcutStyle = style({
   position: 'absolute',
@@ -90,6 +109,11 @@ export const copyLinkShortcutStyle = style({
   opacity: 0.5,
   lineHeight: '20px',
   color: cssVarV2('text/pureWhite'),
+  selectors: {
+    '&.secondary': {
+      color: cssVarV2('text/secondary'),
+    },
+  },
 });
 export const copyLinkTriggerStyle = style({
   padding: '4px 12px 4px 8px',
@@ -109,9 +133,23 @@ export const copyLinkTriggerStyle = style({
     width: '1px',
     backgroundColor: cssVarV2('button/innerBlackBorder'),
   },
+  selectors: {
+    '&.secondary': {
+      backgroundColor: cssVarV2('button/secondary'),
+      color: cssVarV2('text/secondary'),
+    },
+    '&.secondary:hover': {
+      backgroundColor: cssVarV2('button/secondary'),
+      color: cssVarV2('text/secondary'),
+    },
+  },
 });
 globalStyle(`${copyLinkTriggerStyle} svg`, {
   color: cssVarV2('button/pureWhiteText'),
+  transform: 'translateX(2px)',
+});
+globalStyle(`${copyLinkTriggerStyle}.secondary svg`, {
+  color: cssVarV2('text/secondary'),
   transform: 'translateX(2px)',
 });
 export const copyLinkMenuItemStyle = style({
